requirements.txt
fastapi
supabase
pydantic
google.generativeai
dotenv
uvicorn